requirements.txt
pytest~=8.3.5
selene~=2.0.0rc9
selenium~=4.31.0
allure-pytest
allure-python-commons~=2.14.0
python-dotenv~=1.1.0
attach~=0.1